fix: host-app addQuestion util로 이동
useSocket 내부에 정의되어있던
question 을 새로 생성하는 함수를 util.js 로 뺌
